Move Path-Painting Operators to PathPainting folder
