+++ src/UglyToad.PdfPig/Graphics/Operations/PathPainting/CloseFillPathNonZeroWindingAndStroke.cs
@@ -1,34 +1,34 @@
-﻿namespace UglyToad.PdfPig.Graphics.Operations
+﻿namespace UglyToad.PdfPig.Graphics.Operations.PathPainting
 {
     using System.IO;
-
+    
     /// <inheritdoc />
     /// <summary>
-    /// Fill and then stroke the path, using the even-odd rule to determine the region to fill.
+    /// Close, fill, and then stroke the path, using the nonzero winding number rule to determine the region to fill. 
     /// </summary>
-    public class FillPathEvenOddRuleAndStroke : IGraphicsStateOperation
+    public class CloseFillPathNonZeroWindingAndStroke : IGraphicsStateOperation
     {
         /// <summary>
         /// The symbol for this operation in a stream.
         /// </summary>
-        public const string Symbol = "B*";
-
+        public const string Symbol = "b";
+        
         /// <summary>
-        /// The instance of the <see cref="FillPathEvenOddRuleAndStroke"/> operation.
+        /// The instance of the <see cref="CloseFillPathNonZeroWindingAndStroke"/> operation.
         /// </summary>
-        public static readonly FillPathEvenOddRuleAndStroke Value = new FillPathEvenOddRuleAndStroke();
+        public static readonly CloseFillPathNonZeroWindingAndStroke Value = new CloseFillPathNonZeroWindingAndStroke();
 
         /// <inheritdoc />
         public string Operator => Symbol;
 
-        private FillPathEvenOddRuleAndStroke()
-        {
+        private CloseFillPathNonZeroWindingAndStroke()
+        {            
         }
 
         /// <inheritdoc />
         public void Run(IOperationContext operationContext)
         {
-            operationContext.FillPath(false);
+            operationContext.FillPath(true);
         }
 
         /// <inheritdoc />
