+++ src/UglyToad.PdfPig/Graphics/Operations/PathPainting/StrokePath.cs
@@ -1,4 +1,4 @@
-﻿namespace UglyToad.PdfPig.Graphics.Operations
+﻿namespace UglyToad.PdfPig.Graphics.Operations.PathPainting
 {
     using System.IO;
