+++ src/UglyToad.PdfPig/Graphics/ReflectionGraphicsStateOperationFactory.cs
@@ -1,9 +1,5 @@
 namespace UglyToad.PdfPig.Graphics
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Reflection;
     using Operations;
     using Operations.ClippingPaths;
     using Operations.Compatibility;
@@ -17,7 +13,12 @@ namespace UglyToad.PdfPig.Graphics
     using Operations.TextShowing;
     using Operations.TextState;
     using PdfPig.Core;
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Reflection;
     using Tokens;
+    using UglyToad.PdfPig.Graphics.Operations.PathPainting;
 
     internal class ReflectionGraphicsStateOperationFactory : IGraphicsStateOperationFactory
     {
